+++ user_management/api/tests/urls.py
@@ -2,7 +2,12 @@ from django.conf.urls import include, patterns, url
 
 
 urlpatterns = patterns('',
-    url(r'', include('user_management.api.urls')),
-    url(r'', include('user_management.api.urls.users')),
-    url(r'', include('user_management.api.urls.verify_email')),
+    url(r'', include(
+        patterns('',
+            url(r'', include('user_management.api.urls')),
+            url(r'', include('user_management.api.urls.users')),
+            url(r'', include('user_management.api.urls.verify_email')),
+        ),
+        namespace='user_management'
+    ))
 )